add new srv ApplyPlanningScene
This service takes a PlanningScene message
and applies it to the monitored scene.

Ideally it should include a `bool success` field,
but it is not possible to apply the scene and check for success
without ABI changes, so leave it out for now. To get this change
pushed to indigo.